+++ changelog.md
@@ -40,6 +40,7 @@ Date format is `MM/DD/YYYY`.
 - getFriendGroups() method to get a list of groups for a users friends
 - getUsersRoles() method to get a list of roles for a user
 - changeGroupOwner() method to change a group's owner
+- claimGroupOwnership() method to claim ownership of a group
 
 ### Fixed